requirements.txt
aiohttp==3.9.3
beautifulsoup4==4.12.3
redis==5.0.1
kafka-python==2.0.2
elasticsearch==8.11.1
pymongo==4.6.1
pydantic==2.6.1
fastapi==0.109.2
uvicorn==0.27.1
python-dotenv==1.0.1
requests==2.31.0
lxml==5.1.0
tqdm==4.66.1
prometheus-client==0.19.0
robotexclusionrulesparser==1.7.1
bloom-filter2==3.0.0 